feat: perfis tributarios, lotes/validade, NFe, contingencia NFC-e e melhorias fiscais
- Perfis tributarios (CRT/CST/CSOSN por produto) e departamentos de preparacao
- Controle de lotes com validade (FEFO) e inventario de estoque
- Pagina NFe (modelo 55) com finalidade, transporte e devolucao no bridge
- Contingencia NFC-e e salvamento local de XMLs (C:/ExodoNFCe)
- Regime normal (CRT 3) com ICMS completo + IBS/CBS (NT 2025.002)
- Tela de monitor de sincronizacao, backup/restore e impressao
- Migracoes SQL: lotes_produto, nfes, colunas de caixa/estoque/produtos
- Bridge 3.5.4: validade do certificado, override de CRT, QR Code 5 parametros

+++ lib/services/grupo_service.dart
@@ -51,7 +51,7 @@ class GrupoService {
   static List<Produto> atualizarEstoquePorGrupo(
     List<Produto> produtos,
     String grupo,
-    int novoEstoque, {
+    double novoEstoque, {
     bool adicionar = false,
   }) {
     return produtos.map((p) {
